Remove instrument support for SMAP metadata
SMAP metadata does not adhere to the `eos` schema with regard to the
instrument information. The `EOS_Instrument` element is missing, which
resulted in the metadata aggregator not extracting the instrument short
name. This in turn resulted in the instrument short name being set to
an empty string that caused CMR validation failure.

This fix removes the code responsible for retrieving and setting the
instrument short name for SMAP metadata.