Refactor handling of SMAP instrument information
Current SMAP instrument metadata is not compatible with the expected
UMM-G mappings. This results in CMR flagging the UMM-G metadata as
invalid. To work around this issue, instrument information is not
carried over to UMM-G. This commit is a cleanup of the first pass
on addressing this issue. Handling of instrument information is now
conditional based on whether any useful data was extracted from the
source metadata so that the functionality remains available should
future SMAP metadata adhere to published UMM-G mappings.