Don't add `Platforms` data for SMAP
There are some mismatches in platform name for SMAP between the
CMR collection definitions and the ISO SMAP metadata. Rather than
try to correct the mismatch for an unknown number of collections
the platform information will be left out of the UMM-G. Previously
the platform name was not included, so nothing is being lost
relative to published metadata.